+++ packages/typoas-react-query/src/index.ts
@@ -2,3 +2,4 @@ export * from './api-context';
 export * from './infinite-query-factory';
 export * from './mutation-factory';
 export * from './query-factory';
+export { getQueryFunctionKey } from './func-names';