+++ packages/typoas-react-query/src/infinite-query-factory.ts
@@ -17,7 +17,7 @@ import {
   SuccessfulStatus,
 } from '@typoas/runtime';
 import { useApiContext } from './api-context';
-import { getUniqueFunctionName } from './func-names';
+import { getQueryFunctionKey } from './func-names';
 import { TypoasFuncStatusType, TypoasReturnType } from './types';
 
 /**
@@ -108,7 +108,7 @@ export function createInfiniteQueryHook<
   InfiniteData<TypoasReturnType<Func, S>, TPageParam>,
   TError
 > {
-  const name = getUniqueFunctionName(func);
+  const name = getQueryFunctionKey(func);
 
   return (options, queryClient) => {
     const { context: localContext } = useApiContext();