+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ---
+
 ## 4.0.0 - 2024-08-21
 
 > All packages were updated to 4.0.0